// src/gen/api/types/ListIncomingFriendRequests.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { FriendRequest } from './FriendRequest.ts'

export type ListIncomingFriendRequestsQueryParams = {
  /**
   * @type string | undefined
   */
  cursor?: string
  /**
   * @minLength 1
   * @maxLength 50
   * @default 20
   * @type integer | undefined
   */
  limit?: number
}

/**
 * @description Cursor-paginated incoming pending requests
 */
export type ListIncomingFriendRequests200 = {
  /**
   * @type array
   */
  data: FriendRequest[]
  /**
   * @type object
   */
  pagination: {
    /**
     * @type string,null
     */
    nextCursor: string | null
    /**
     * @type boolean
     */
    hasNext: boolean
    /**
     * @type number
     */
    limit: number
  }
}

/**
 * @description Validation failed
 */
export type ListIncomingFriendRequests400 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Missing or invalid access token
 */
export type ListIncomingFriendRequests401 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

/**
 * @description Internal server error
 */
export type ListIncomingFriendRequests500 = {
  /**
   * @type string
   */
  error: string
  /**
   * @type string
   */
  message: string
  /**
   * @type string
   */
  code: string
  /**
   * @type string | undefined
   */
  requestId?: string
  /**
   * @type object | undefined
   */
  details?: {
    [key: string]: string
  }
}

export type ListIncomingFriendRequestsQueryResponse =
  ListIncomingFriendRequests200

export type ListIncomingFriendRequestsQuery = {
  Response: ListIncomingFriendRequests200
  QueryParams: ListIncomingFriendRequestsQueryParams
  Errors:
    | ListIncomingFriendRequests400
    | ListIncomingFriendRequests401
    | ListIncomingFriendRequests500
}